+++ src/tools.rs
@@ -689,9 +689,7 @@ mod tests {
     #![allow(clippy::indexing_slicing)]
 
     use super::*;
-    use crate::{
-        config::Config, message::get_msg_info, receive_imf::receive_imf, test_utils::TestContext,
-    };
+    use crate::{message::get_msg_info, receive_imf::receive_imf, test_utils::TestContext};
 
     #[test]
     fn test_parse_receive_headers() {
@@ -762,7 +760,6 @@ DKIM Results: Passed=true, Works=true, Allow_Keychange=true";
 
     async fn check_parse_receive_headers_integration(raw: &[u8], expected: &str) {
         let t = TestContext::new_alice().await;
-        t.set_config(Config::ShowEmails, Some("2")).await.unwrap();
         receive_imf(&t, raw, false).await.unwrap();
         let msg = t.get_last_msg().await;
         let msg_info = get_msg_info(&t, msg.id).await.unwrap();